requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
boto3==1.26.32
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Django==4.0.5
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-timezone-field==5.0
django-utils==0.0.2
djangorestframework==3.13.1
drf-standardized-errors==0.12.4
dnspython==2.2.1
eventlet==0.33.2
flower==1.2.0
greenlet==2.0.1
humanize==4.4.0
kombu==5.2.4
packaging==21.3
prometheus-client==0.15.0
prompt-toolkit==3.0.33
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
redis==4.3.5
six==1.16.0
sqlparse==0.4.2
tornado==6.2
tzdata==2022.6
vine==5.0.0
wcwidth==0.2.5
yapf==0.32.0
requests==2.28.0
Js2Py==0.71
django-extensions==3.2.1
psycopg2-binary==2.9.5
pyjwt==2.6.0
uvicorn==0.20.0
gunicorn==20.1.0